fix: store token ids as uint32 for llm-jp vocab >65536

+++ tests/test_data.py
@@ -7,9 +7,9 @@ def test_sample_batch_shape_and_range(tmp_path):
 
     # 100 tokens の tiny memmap を作成
     bin_path = tmp_path / "tiny.bin"
-    arr = np.arange(100, dtype=np.uint16)
+    arr = np.arange(100, dtype=np.uint32)
     arr.tofile(bin_path)
-    mm = np.memmap(bin_path, dtype=np.uint16, mode="r", shape=(100,))
+    mm = np.memmap(bin_path, dtype=np.uint32, mode="r", shape=(100,))
 
     rng = np.random.default_rng(seed=42)
     x, y = sample_batch(mm, batch_size=4, seq_len=8, rng=rng)
@@ -28,9 +28,9 @@ def test_sample_batch_y_is_x_shifted_by_one(tmp_path):
     from openmythos_pg.data import sample_batch
 
     bin_path = tmp_path / "seq.bin"
-    arr = np.arange(200, dtype=np.uint16)
+    arr = np.arange(200, dtype=np.uint32)
     arr.tofile(bin_path)
-    mm = np.memmap(bin_path, dtype=np.uint16, mode="r", shape=(200,))
+    mm = np.memmap(bin_path, dtype=np.uint32, mode="r", shape=(200,))
 
     rng = np.random.default_rng(seed=0)
     x, y = sample_batch(mm, batch_size=2, seq_len=16, rng=rng)
@@ -43,9 +43,9 @@ def test_sample_batch_reproducible_with_same_seed(tmp_path):
     from openmythos_pg.data import sample_batch
 
     bin_path = tmp_path / "seq.bin"
-    arr = np.arange(500, dtype=np.uint16)
+    arr = np.arange(500, dtype=np.uint32)
     arr.tofile(bin_path)
-    mm = np.memmap(bin_path, dtype=np.uint16, mode="r", shape=(500,))
+    mm = np.memmap(bin_path, dtype=np.uint32, mode="r", shape=(500,))
 
     x1, _ = sample_batch(mm, 4, 16, np.random.default_rng(seed=123))
     x2, _ = sample_batch(mm, 4, 16, np.random.default_rng(seed=123))
